Catuserbot Version -2.9.2
Complete info is done you can check by .info
 ☞ Updated python to 3.9.0
 ☞ Custom alive header and alive emoji's added
 ☞ CUSTOM_STICKER_PACKNAME  is added
 ☞ OWNER_ID is added to set the required account as owner
 ☞ Added pin support in pm too and also unpin option for both groups and pm
 ☞ For stoping autoprofile commands instead of restarting bot added end option by command
 ☞ dice_dart_ball is renamed to emojigames and slot machine is added
 ☞ Small changes in evaluators
 ☞ removed extdl and it install automatically after every restart
 ☞ small changes in scam
 ☞ ffmpeg is modified like that you can extract audio only and trim or video trim
 ☞ Added file summary option counts no of files either for user or in chat . thanks to @deleteduser420 and @spechide
 ☞ fpost is fixed
 ☞ Pmpermit is modified and added variables like {mention} , {username} for CUSTOM_PMPERMIT_TEXT and also without mentioning user you can aprove user from any where by his username
 ☞ fixed webp format for remove background
 ☞ Translater is fixed
 ☞ Finally some minor bugs fixes

+++ userbot/plugins/forward.py
@@ -1,9 +1,25 @@
 import string
 
+from telethon.tl.types import Channel
+
 from ..utils import admin_cmd, edit_or_reply, sudo_cmd
 from . import CMD_HELP
 
+global msg_cache
 msg_cache = {}
+
+
+global groupsid
+groupsid = []
+
+
+async def all_groups_id(cat):
+    catgroups = []
+    async for dialog in cat.client.iter_dialogs():
+        entity = dialog.entity
+        if isinstance(entity, Channel) and entity.megagroup:
+            catgroups.append(entity.id)
+    return catgroups
 
 
 @bot.on(admin_cmd(pattern="frwd$"))
@@ -32,6 +48,8 @@ async def _(event):
 @bot.on(admin_cmd(pattern="resend$"))
 @bot.on(sudo_cmd(pattern="resend$", allow_sudo=True))
 async def _(event):
+    if event.fwd_from:
+        return
     try:
         await event.delete()
     except:
@@ -42,31 +60,43 @@ async def _(event):
     await event.respond(m)
 
 
-@bot.on(admin_cmd(pattern=r"fpost\s+(.*)"))
-@bot.on(sudo_cmd(pattern="fpost\s+(.*)", allow_sudo=True))
+@bot.on(admin_cmd(pattern=r"fpost (.*)"))
+@bot.on(sudo_cmd(pattern=r"fpost (.*)", allow_sudo=True))
 async def _(event):
+    if event.fwd_from:
+        return
+    global groupsid
+    global msg_cache
     await event.delete()
     text = event.pattern_match.group(1)
     destination = await event.get_input_chat()
+    if len(groupsid) == 0:
+        groupsid = await all_groups_id(event)
     for c in text.lower():
         if c not in string.ascii_lowercase:
             continue
         if c not in msg_cache:
-            async for msg in event.client.iter_messages(None, search=c):
+            async for msg in event.client.iter_messages(event.chat_id, search=c):
                 if msg.raw_text.lower() == c and msg.media is None:
                     msg_cache[c] = msg
                     break
+        if c not in msg_cache:
+            for i in groupsid:
+                async for msg in event.client.iter_messages(event.chat_id, search=c):
+                    if msg.raw_text.lower() == c and msg.media is None:
+                        msg_cache[c] = msg
+                        break
         await event.client.forward_messages(destination, msg_cache[c])
 
 
 CMD_HELP.update(
     {
         "forward": "**Plugin : **`forward`\
-    \n\n**Synatax : **`frwd reply to any message`\
-    \n**Function :  **__Enable Seen Counter in any message, to know how many users have seen your message__\
-    \n\n**Syntax : **`.resend reply to message`\
-    \n**Function : **__Just resend the replied message again in that chat__\
-    \n\n**Syntax : **`.fpost text`\
-    \n**Function : **__Split the word and forwards each letter from the messages cache if exists__"
+    \n\n  •  **Synatax : **`frwd reply to any message`\
+    \n  •  **Function :  **__Enable Seen Counter in any message, to know how many users have seen your message__\
+    \n\n  •  **Syntax : **`.resend reply to message`\
+    \n  •  **Function : **__Just resend the replied message again in that chat__\
+    \n\n  •  **Syntax : **`.fpost text`\
+    \n  •  **Function : **__Split the word and forwards each letter from the messages cache if exists__"
     }
 )
